[ARV-223] rename: test codes integrations are devided application and query
